@Romeli 添加一些注释 错误和命名

+++ Tool/Convert.cpp
@@ -10,6 +10,14 @@
 namespace User {
 namespace Tool {
 
+/*
+ * author Romeli
+ * explain 将整形数转换为字符串
+ * param1 num 整形数
+ * param2 base 进制
+ * param3 str 目标字符串地址
+ * return uint8_t
+ */
 uint8_t Convert::byNumber(int32_t num, uint8_t base, uint8_t* str) {
 	uint8_t len = 0;
 	//小于零取反加负号
@@ -31,6 +39,14 @@ uint8_t Convert::byNumber(int32_t num, uint8_t base, uint8_t* str) {
 	return len;
 }
 
+/*
+ * author Romeli
+ * explain 将浮点数转换为字符串
+ * param1 flo 浮点数
+ * param2 ndigit 小数位数
+ * param3 str 目标字符串地址
+ * return uint8_t
+ */
 uint8_t Convert::byFloat(double flo, uint8_t ndigit, uint8_t* str) {
 	uint8_t len = 0, len2 = 0;
 	uint32_t int_part = 0;
@@ -71,6 +87,13 @@ uint8_t Convert::byFloat(double flo, uint8_t ndigit, uint8_t* str) {
 	return len;
 }
 
+/*
+ * author Romeli
+ * explain 根据进制计算整形数长度
+ * param1 num 整形数
+ * param2 base 进制
+ * return uint8_t
+ */
 uint8_t Convert::getLen(uint32_t num, uint8_t base) {
 	uint8_t i;
 	if (num == 0) {
@@ -82,13 +105,28 @@ uint8_t Convert::getLen(uint32_t num, uint8_t base) {
 	return i;
 }
 
-double Convert::pow10(uint8_t power) {
+/*
+ * author Romeli
+ * explain 求10的n次方
+ * param power 次方数
+ * return uint32_t
+ */
+uint32_t Convert::pow10(uint8_t power) {
 	if (power == 0)
-		return 1.0;
+		return 1;
 	else
-		return 10.0 * pow10(--power);
+		return 10 * pow10(--power);
 }
 
+/*
+ * author Romeli
+ * explain 两个字符串拼接
+ * param1 str_to 目标字符串
+ * param2 str_to_len 目标字符串长度
+ * param3 str_from 被拼接的字符串
+ * param4 str_from_len 被拼接的字符串长度
+ * return uint8_t
+ */
 uint8_t Convert::strcat(uint8_t* str_to, uint8_t str_to_len, uint8_t* str_from,
 		uint8_t str_from_len) {
 	uint8_t i;
